% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overDueRpt.R
\name{saleOverdueRptUI}
\alias{saleOverdueRptUI}
\title{用户新增UI}
\usage{
saleOverdueRptUI(
  tabTitle = "业务应收报表查询",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(4, 8),
  func_left = saleOverdueRptFunc_left,
  func_right = saleOverdueRptFunc_right
)
}
\arguments{
\item{tabTitle}{标签名称}

\item{colTitles}{列名}

\item{widthRates}{比例}

\item{func_left}{函数}

\item{func_right}{函数}
}
\value{
返回值
}
\description{
用户新增UI
}
\examples{
saleOverdueRptUI()
}
